Refactor: ImageDetailModal과 ImageMenuBar 병합
신고, 제거 핸들러 함수 이름 변경
imageDetails에 있는 imageId 사용
